refactor(internship_offer): start to split logic based with weekly_framed
