+++ app/models/internship_offer_week.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 class InternshipOfferWeek < ApplicationRecord
-  belongs_to :internship_offer, counter_cache: true
+  belongs_to :internship_offer, counter_cache: true#,
+                                #inverse_of: :internship_offer_weeks
   belongs_to :week
 
   has_many :internship_applications, dependent: :destroy
