+++ app/views/dashboard/internship_offers/_form.html.slim
@@ -88,7 +88,6 @@
           | ; utilisez des mots simples. Qu'est-ce que l'élève va découvrir (ou faire) pendant sa semaine de stage ? Donnez 3 qualités requises (aimer bricoler, être bavard, aimer écrire ...).
           = " "
           = link_to "Besoin d'aide à la rédaction ?", les_10_commandements_d_une_bonne_offre_path, title: "Consulter les 10 commandements d'une bonne offre de stage (nouvelle fenêtre)", target: "_blank", rel: "external noopener noreferrer"
-
       = form.rich_text_area :description_rich_text, class: "form-control col-lg-12", required: true, data: { target: 'maxlength-input.trixElement' }
       small.form-text data-target="maxlength-input.trixElementCharCount"
